# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Course Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all published courses
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseListResponse"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a new course (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseBody"
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: role
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /admin/users/{userId}:
    get:
      operationId: adminGetUser
      tags: [admin]
      summary: Get user details (admin)
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
    put:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update user (admin)
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{userId}/ban:
    post:
      operationId: banUser
      tags: [admin]
      summary: Ban/unban user (admin)
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanUserBody"
      responses:
        "200":
          description: User ban status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/revenue:
    get:
      operationId: getRevenueReport
      tags: [admin]
      summary: Get revenue report with chart data
      parameters:
        - in: query
          name: period
          schema:
            type: string
            enum: [7d, 30d, 90d, 1y]
            default: 30d
      responses:
        "200":
          description: Revenue report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueReport"
...
  /admin/payouts/{payoutId}/approve:
    post:
      operationId: approvePayout
      tags: [admin]
      summary: Approve a payout request (admin)
      parameters:
        - in: path
          name: payoutId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payout approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/payouts/{payoutId}/reject:
    post:
      operationId: rejectPayout
      tags: [admin]
      summary: Reject a payout request (admin)
      parameters:
        - in: path
          name: payoutId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectPayoutBody"
      responses:
        "200":
          description: Payout rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/enrollments:
    get:
      operationId: adminListEnrollments
      tags: [admin]
      summary: List all enrollments (admin)
      parameters:
        - in: query
          name: courseId
          schema:
            type: integer
        - in: query
          name: userId
          schema:
            type: integer
      responses:
        "200":
          description: Enrollment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminEnrollment"
...
  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin]
      summary: Get platform settings (admin)
      responses:
        "200":
          description: Platform settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"
    put:
      operationId: updateAdminSettings
      tags: [admin]
      summary: Update platform settings (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsBody"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"

  # Coupons
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # Auth schemas
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, student, affiliate]
        avatarUrl:
          type: string
          nullable: true
        referralCode:
          type: string
        isBanned:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, email, name, role, referralCode, isBanned, createdAt]
    UserDetail:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            enrollmentCount:
              type: integer
            totalSpent:
              type: number
            affiliateEarnings:
              type: number
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required: [users, total, limit, offset]

    # Course schemas
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        thumbnailUrl:
          type: string
          nullable: true
        price:
          type: number
        category:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        status:
          type: string
          enum: [draft, published]
        enrollmentCount:
          type: integer
        moduleCount:
          type: integer
        lessonCount:
          type: integer
        durationMinutes:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, description, price, category, level, status, enrollmentCount, moduleCount, lessonCount, durationMinutes, createdAt, updatedAt]
...
    CourseListResponse:
      type: object
      properties:
        courses:
          type: array
          items:
            $ref: "#/components/schemas/Course"
        total:
          type: integer
      required: [courses, total]
    CreateCourseBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        thumbnailUrl:
          type: string
          nullable: true
        price:
          type: number
        category:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        status:
          type: string
          enum: [draft, published]
      required: [title, description, price, category, level]
...
    AdminEnrollment:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        userId:
          type: integer
        enrolledAt:
          type: string
          format: date-time
        progressPercent:
          type: number
        user:
          $ref: "#/components/schemas/User"
        course:
          $ref: "#/components/schemas/Course"
      required: [id, courseId, userId, enrolledAt, progressPercent]
...
    RejectPayoutBody:
      type: object
      properties:
        reason:
          type: string
      required: [reason]

    # Admin analytics
...
    RevenueReport:
      type: object
      properties:
        period:
          type: string
        totalRevenue:
          type: number
        chartData:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
        byGateway:
          type: object
          properties:
            stripe:
              type: number
            razorpay:
              type: number
          required: [stripe, razorpay]
      required: [period, totalRevenue, chartData, byGateway]
    ChartDataPoint:
      type: object
      properties:
        date:
          type: string
        revenue:
          type: number
        enrollments:
          type: integer
      required: [date, revenue, enrollments]

    # Platform settings
    PlatformSettings:
      type: object
      properties:
        siteName:
          type: string
        siteDescription:
          type: string
        commissionRate:
          type: number
        currency:
          type: string
        stripeEnabled:
          type: boolean
        razorpayEnabled:
          type: boolean
        emailNotificationsEnabled:
          type: boolean
      required: [siteName, siteDescription, commissionRate, currency, stripeEnabled, razorpayEnabled, emailNotificationsEnabled]
    UpdateSettingsBody:
      type: object
      properties:
        siteName:
          type: string
        siteDescription:
          type: string
        commissionRate:
          type: number
        currency:
          type: string
        stripeEnabled:
          type: boolean
        razorpayEnabled:
          type: boolean
        emailNotificationsEnabled:
          type: boolean

    # Admin user update
    AdminUpdateUserBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [admin, student, affiliate]
        isBanned:
          type: boolean
    BanUserBody:
      type: object
      properties:
        banned:
          type: boolean
      required: [banned]

    # Coupons
...